// src/app/modules/usuario/models/docenteModel.ts
export interface DocenteModel {
    nombre_usuario: string;
    contrasena_usuario: string;
    rol_usuario: number;
    nombres_docente: string;
    apellidos_docente: string;
    cedula_docente: string;
    fechaNacimiento_docente: Date;
    edad_docente: number;
    direccion_docente: string;
    telefono_docente: string;
    email_docente: string;
    titulo_docente: string;
    nivelEducacion_docente: string;
    promedio_docente: number;
  }
  